Rename HNIrrigationCriteria.* to HNIrrigationPlacement.*
