+++ CMakeLists.txt
@@ -28,7 +28,7 @@ SET( HNIRRIGATIOND_SRC
      ${CMAKE_SOURCE_DIR}/src/daemon/HNSWDPacketClient.cpp
      ${CMAKE_SOURCE_DIR}/src/daemon/HNSWDStatus.cpp
      ${CMAKE_SOURCE_DIR}/src/daemon/HNIrrigationZone.cpp
-     ${CMAKE_SOURCE_DIR}/src/daemon/HNIrrigationCriteria.cpp
+     ${CMAKE_SOURCE_DIR}/src/daemon/HNIrrigationPlacement.cpp
      ${CMAKE_SOURCE_DIR}/src/daemon/HNIrrigationTypes.cpp
 )
 
@@ -36,7 +36,7 @@ SET( HNI_TEST_SCHEDULER_SRC
      ${CMAKE_SOURCE_DIR}/src/daemon/hni_test_scheduler.cpp
      ${CMAKE_SOURCE_DIR}/src/daemon/HNIrrigationSchedule.cpp
      ${CMAKE_SOURCE_DIR}/src/daemon/HNIrrigationZone.cpp
-     ${CMAKE_SOURCE_DIR}/src/daemon/HNIrrigationCriteria.cpp
+     ${CMAKE_SOURCE_DIR}/src/daemon/HNIrrigationPlacement.cpp
      ${CMAKE_SOURCE_DIR}/src/daemon/HNIrrigationTypes.cpp
 )
