+++ src/daemon/hni_test_scheduler.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
 
 #include "HNIrrigationZone.h"
-#include "HNIrrigationCriteria.h"
+#include "HNIrrigationPlacement.h"
 #include "HNIrrigationSchedule.h"
 
 class HNIrrigationTest
